tenant_id checking reuse current controller logic
complements fixing THREESCALE-10686
